Do not reset preloaded collections with decorate
When `all` is called, it instantiates a new collection proxy, which will
not be loaded. As a result, this breaks preloading.

This patch fixes it by implementing the `decorate` method on
`CollectionProxy`, where we're able to access `#self` (as opposed to
having to call `.all`) in order to access the relation.